+++ src/frontend/src/components/RegForm/RegForm.jsx
@@ -4,7 +4,7 @@ import { useNavigate } from "react-router-dom";
 import useUserStore from "../../actions/useUserStore";
 import NotificationBox from "../NotificationBox/NotificationBox";
 import { axiosInstance } from "../../utils/axiosInstance";
-import API from "../../utils/API/TEST_API";
+import API from "../../utils/API/API";
 
 function RegForm() {
   const setUserInfo = useUserStore((state) => state.setUserInfo);
