+++ src/frontend/src/utils/API/API.js
@@ -1,29 +1,34 @@
 const BASE_URL = "https://fittrip.site/api";
-const CHAT_API_BASE_URL = "http://localhost:7070";
 const COMMUNITY_BASE_URL = "https://fittrip.site/api/community";
 const MEDIA_BASE_URL = "https://fittrip.site/api/sig";
-const NOTIFICATION_URL = "https://fittrip.site/api/notification"
-const WS_SERVER_BASE = "ws/api/chat/server/message";
-const WS_DM_BASE = "ws/api/chat/direct/message";
-const WS_FORUM_BASE = "/ws/api/chat/forum/message";
+const NOTIFICATION_URL = "https://fittrip.site/api/notification";
+const WS_BASE = "ws/api/chat";
 const API = {
+  //유저
+  LOG_IN: `${BASE_URL}/user/login`,
+  SIGN_UP: `${BASE_URL}/user/signup`,
+  LOG_OUT: `${BASE_URL}/user/logout`,
+  GET_USER_PROFLIE: `${BASE_URL}/user/profile`,
+
+  //커뮤니티 회원가입
+  COMM_SIGNUP: `${BASE_URL}/community/user`,
+  READ_MAIN: (userId) => `${BASE_URL}/community/user/${userId}`,
+
   //서버(커뮤니티) 채팅
   SUBSCRIBE_CHAT: (serverId) => `/topic/server/${serverId}`,
-  SEND_CHAT: `${WS_SERVER_BASE}/send`,
-  MODIFY_CHAT: `${WS_SERVER_BASE}/modify`,
-  DELETE_CHAT: `${WS_SERVER_BASE}/delete`,
-  IS_TYPING: `${WS_SERVER_BASE}/typing`,
-  FILE_UPLOAD: `${CHAT_API_BASE_URL}/server/message/file`,
-  GET_HISTORY: `${CHAT_API_BASE_URL}/server/messages/channel`,
-  GET_COMMENTS: `${CHAT_API_BASE_URL}/server/comments/message`,
-  POST_LOCATION: `${CHAT_API_BASE_URL}/server/user/location`,
+  FILE_UPLOAD: `${BASE_URL}/chat/server/message/file`,
+  GET_HISTORY: `${BASE_URL}/chat/server/messages/channel`,
+  GET_COMMENTS: `${BASE_URL}/chat/server/comments/message`,
+  POST_LOCATION: `${BASE_URL}/chat/server/user/location`,
 
   //서버 이벤트
-  COMM_SERVER: `${COMMUNITY_BASE_URL}/server`, 
-  GET_SERVER: (serverId, userId) =>`${COMMUNITY_BASE_URL}/server/${serverId}/${userId}`,
-  INVITE_SERVER: (serverId) => `${COMMUNITY_BASE_URL}/server/${serverId}/invitation`,
+  COMM_SERVER: `${COMMUNITY_BASE_URL}/server`,
+  GET_SERVER: (serverId, userId) =>
+    `${COMMUNITY_BASE_URL}/server/${serverId}/${userId}`,
+  INVITE_SERVER: (serverId) =>
+    `${COMMUNITY_BASE_URL}/server/${serverId}/invitation`,
   JOIN_SERVER: `${COMMUNITY_BASE_URL}/server/join`,
-  COMM_CATEGORY:`${COMMUNITY_BASE_URL}/category`,
+  COMM_CATEGORY: `${COMMUNITY_BASE_URL}/category`,
   COMM_CHANNEL: `${COMMUNITY_BASE_URL}/channel`,
 
   //미디어
@@ -35,26 +40,25 @@ const API = {
 
   //DM 채팅
   SUBSCRIBE_DM: (roomId) => `/topic/direct/${roomId}`,
-  SEND_DM: `${WS_DM_BASE}/send`,
-  MODIFY_DM: `${WS_DM_BASE}/modify`,
-  DELETE_DM: `${WS_DM_BASE}/delete`,
-  IS_TYPING_DM: `${WS_DM_BASE}/typing`,
-  DM_FILE_UPLOAD: `${CHAT_API_BASE_URL}/direct/message/file`,
-  GET_DM_HISTORY: `${CHAT_API_BASE_URL}/direct/messages/channel`,
-  GET_DM_COMMENTS: `${CHAT_API_BASE_URL}/direct/comments`,
-
-  //DM이벤트
+  DM_FILE_UPLOAD: `${BASE_URL}/chat/direct/message/file`,
+  GET_DM_HISTORY: `${BASE_URL}/chat/direct/messages/channel`,
+  GET_DM_COMMENTS: `${BASE_URL}/chat/direct/comments`,
 
   //FORUM 채팅
-  SEND_FORUM: `${WS_FORUM_BASE}/send`,
-  MODIFY_FORUM: `${WS_FORUM_BASE}/modify`,
-  DELETE_FORUM: `${WS_FORUM_BASE}/delete`,
-  IS_TYPING_FORUM: `${WS_FORUM_BASE}/typing`,
-  FORUM_FILE_UPLOAD: `${CHAT_API_BASE_URL}/forum/message/file`,
-  GET_FORUM_HISTORY: `${CHAT_API_BASE_URL}/forum/messages/forum`,
-  GET_FORUM_COMMENTS: `${CHAT_API_BASE_URL}/forum/comments`,
+  FORUM_FILE_UPLOAD: `${BASE_URL}/chat/forum/message/file`,
+  GET_FORUM_HISTORY: `${BASE_URL}/chat/forum/messages/forum`,
+  GET_FORUM_COMMENTS: `${BASE_URL}/chat/forum/comments`,
 
   //FORUM 이벤트
+  READ_FORUM: (channelId, userId) =>
+    `${BASE_URL}/community/channel/${channelId}/${userId}`,
+  CUD_FORUM: `${BASE_URL}/community/forum`,
+
+  //소켓
+  SEND_CHAT: (TYPE) => `${WS_BASE}/${TYPE}/message/send`,
+  DELETE_CHAT: (TYPE) => `${WS_BASE}/${TYPE}/message/delete`,
+  MODIFY_CHAT: (TYPE) => `${WS_BASE}/${TYPE}/message/modify`,
+  ISTYPING: (TYPE) => `${WS_BASE}/${TYPE}/message/typing`,
 };
 
 export default API;
